refactor: remote notify command priority
