+++ app/src/main/java/uk/nktnet/webviewkiosk/config/remote/RemoteNotifyCommandPriority.kt
@@ -1,4 +1,4 @@
-package uk.nktnet.webviewkiosk.config.mqtt
+package uk.nktnet.webviewkiosk.config.remote
 
 import androidx.core.app.NotificationCompat
 import kotlinx.serialization.KSerializer
@@ -9,8 +9,8 @@ import kotlinx.serialization.descriptors.SerialDescriptor
 import kotlinx.serialization.encoding.Decoder
 import kotlinx.serialization.encoding.Encoder
 
-@Serializable(with = MqttNotifyPrioritySerializer::class)
-enum class MqttNotifyPriority(val androidValue: Int, val label: String) {
+@Serializable(with = RemoteNotifyCommandPrioritySerializer::class)
+enum class RemoteNotifyCommandPriority(val androidValue: Int, val label: String) {
     MIN(NotificationCompat.PRIORITY_MIN, "Min"),
     LOW(NotificationCompat.PRIORITY_LOW, "Low"),
     DEFAULT(NotificationCompat.PRIORITY_DEFAULT, "Default"),
@@ -18,7 +18,7 @@ enum class MqttNotifyPriority(val androidValue: Int, val label: String) {
     MAX(NotificationCompat.PRIORITY_MAX, "Max");
 
     companion object {
-        fun fromString(value: String?): MqttNotifyPriority =
+        fun fromString(value: String?): RemoteNotifyCommandPriority =
             entries.find {
                 it.name.equals(value, ignoreCase = true)
                 || it.label.equals(value, ignoreCase = true)
@@ -27,18 +27,18 @@ enum class MqttNotifyPriority(val androidValue: Int, val label: String) {
     }
 }
 
-object MqttNotifyPrioritySerializer : KSerializer<MqttNotifyPriority> {
+object RemoteNotifyCommandPrioritySerializer : KSerializer<RemoteNotifyCommandPriority> {
     override val descriptor: SerialDescriptor = PrimitiveSerialDescriptor(
-        "MqttNotifyPriority",
+        "RemoteNotifyCommandPrioritySerializer",
         PrimitiveKind.STRING
     )
 
-    override fun serialize(encoder: Encoder, value: MqttNotifyPriority) {
+    override fun serialize(encoder: Encoder, value: RemoteNotifyCommandPriority) {
         encoder.encodeString(value.label)
     }
 
-    override fun deserialize(decoder: Decoder): MqttNotifyPriority {
+    override fun deserialize(decoder: Decoder): RemoteNotifyCommandPriority {
         val str = decoder.decodeString()
-        return MqttNotifyPriority.fromString(str)
+        return RemoteNotifyCommandPriority.fromString(str)
     }
 }
